Add pagination hook
